Move original source file to the /src folder
